+++ package.json
@@ -20,8 +20,7 @@
     "node-sass-middleware": "^0.11.0",
     "nodemailer": "^2.7.2",
     "sanitizer": "^0.1.3",
-    "stripe": "^4.15.0",
-    "stripe-webhook-middleware": "^0.2.0"
+    "stripe": "^4.15.0"
   },
   "devDependencies": {
     "eslint": "^3.14.1",
