+++ templates/views/goProThanks.hbs
@@ -36,6 +36,9 @@
           <p>
             There's more coming! You pro subscription will continue to gain new perks in the coming months as we flesh out the rest of the website. Stay tuned!
           </p>
+          <p>
+            <button type="button" id="testBtn">TEST STRIPE EVENT</button>
+          </p>
         </section>
         <!-- /TEXT -->
 
@@ -49,3 +52,33 @@
         {{> sidebar_twitter }}
       </aside>
       <!-- /SIDEBAR -->
+
+<script>
+
+var stripeJSON = {"created":1326853478,"livemode":false,"id":"evt_00000000000000","type":"charge.succeeded","object":"event","request":null,"pending_webhooks":1,"api_version":"2017-01-27","data":{"object":{"id":"ch_00000000000000","object":"charge","amount":1499,"amount_refunded":0,"application":null,"application_fee":null,"balance_transaction":"txn_00000000000000","captured":true,"created":1487698362,"currency":"usd","customer":"cus_00000000000000","description":null,"destination":null,"dispute":null,"failure_code":null,"failure_message":null,"fraud_details":{},"invoice":"in_00000000000000","livemode":false,"metadata":{},"on_behalf_of":null,"order":null,"outcome":{"network_status":"approved_by_network","reason":null,"risk_level":"normal","seller_message":"Payment complete.","type":"authorized"},"paid":true,"receipt_email":null,"receipt_number":null,"refunded":false,"refunds":{"object":"list","data":[],"has_more":false,"total_count":0,"url":"/v1/charges/ch_19phVGK2sFMaOukMrgqIN6C8/refunds"},"review":null,"shipping":null,"source":{"id":"card_00000000000000","object":"card","address_city":null,"address_country":null,"address_line1":null,"address_line1_check":null,"address_line2":null,"address_state":null,"address_zip":"06598","address_zip_check":"pass","brand":"Visa","country":"US","customer":"cus_00000000000000","cvc_check":"pass","dynamic_last4":null,"exp_month":8,"exp_year":2018,"funding":"credit","last4":"4242","metadata":{},"name":"chris@closebrace.com","tokenization_method":null},"source_transfer":null,"statement_descriptor":null,"status":"succeeded","transfer_group":null}}  };
+
+document.addEventListener('DOMContentLoaded', function() {
+  document.getElementById('testBtn').addEventListener('click', function(e){
+    e.preventDefault();
+    aja()
+      .method('post')
+      .url('/api/stripe/events')
+      .cache(false)
+      .body(stripeJSON)
+      .on('200', function(response){
+        // Reload the window to show the comment
+        console.log('200');
+      })
+       .on('40x', function(response){
+          //something is definitely wrong
+          // 'x' means any number (404, 400, etc. will match)
+          console.log(response);
+      })
+      .on('500', function(response){
+          //oh crap
+          console.log(response);
+      })
+      .go();
+  });
+});
+</script>
